+++ NEWS.md
@@ -2,6 +2,7 @@
 - Paginate dataset summary to prevent `skimr::skim()` freezing GUI when number of variables in the dataset is large
 - New 'GTag' and 'GTags' classes for labels with a background and a collection of droppable labels, respectively
 - add search button/clear button to search group (in data view widget)
+- iNZDocument class uses global option `inzight.default.par` to override the default values for any `iNZightPlots::inzpar()` settings
 
 **Interface changes**